fix(chore): Move NftSection from routes to components
