+++ client/src/app/components/stardust/NftSection.scss
