--- client/src/app/routes/stardust/INftDetails.ts
@@ -1,15 +1,0 @@
-export default interface INftDetails {
-    /**
-     * NFT image.
-     */
-    image?: string;
-    /**
-     * NFT name.
-     */
-    name?: string;
-    /**
-     * NFT id.
-     */
-    id: string;
-}
-
