--- client/src/app/routes/stardust/ITokenDetails.ts
@@ -1,23 +1,0 @@
-export default interface ITokenDetails {
-    /**
-     * Token name.
-     */
-    name: string;
-    /**
-     * Token symbol.
-     */
-    symbol?: string;
-    /**
-     * Token held amount.
-     */
-    amount: number;
-    /**
-     * Token price.
-     */
-    price?: number;
-    /**
-     * Token total value held.
-     */
-    value?: number;
-}
-
